multiple thermostat support
added support for multiple thermostats
