+++ config.schema.json
@@ -4,35 +4,40 @@
   "singular": true,
   "headerDisplay": "<p align=\"center\"><img width=\"30%\" src=\"https://user-images.githubusercontent.com/8211291/131715404-6aa429a1-6a57-447c-b1b6-7ac84ee79977.jpg\"</p>",
   "schema": {
-    "type": "object",
-    "properties": {
-      "url": {
-        "title": "URL of Infinitude e.g. http://{host}:{port}",
-        "type": "string",
-        "format": "uri",
-        "required": true
-      },
-      "holdUntil": {
-        "title": "Time to hold until (will hold forever if not specified), also used when no activity scheduled in t-stat",
-        "type": "string",
-        "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
-      },
-      "shutOffAway": {
-        "title": "Set system to Away when set to Off in HomeKit",
-        "type": "boolean",
-        "default": false
-      },
-      "holdUntilNextActivity": {
-        "title": "Hold until next scheduled activity",
-        "type": "boolean",
-        "default": false
-      },
-      "useOutdoorTemperatureSensor": {
-        "title": "Create outdoor temperature sensor",
-        "type": "boolean",
-        "default": true
+    "thermostats": {
+      "type": "array",
+      "items": {
+        "title": "Thermostats",
+        "type": "object",
+        "properties": {
+          "url": {
+            "title": "URL of Infinitude e.g. http://{host}:{port}",
+            "type": "string",
+            "format": "uri",
+            "required": true
+          },
+          "holdUntil": {
+            "title": "Time to hold until (will hold forever if not specified), also used when no activity scheduled in this thermostat",
+            "type": "string",
+            "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
+          },
+          "shutOffAway": {
+            "title": "Set system to Away when set to Off in HomeKit for this thermostat",
+            "type": "boolean",
+            "default": false
+          },
+          "holdUntilNextActivity": {
+            "title": "Hold until next scheduled activity for this thermostat",
+            "type": "boolean",
+            "default": false
+          },
+          "useOutdoorTemperatureSensor": {
+            "title": "Create outdoor temperature sensor for this thermostat",
+            "type": "boolean",
+            "default": true
+          }
+        }
       }
     }
   }
 }
-
